fix(SearchContactComposite): fix the referenceId to keep only letters, numbers and underscores (#17)
## Description
The customer wants to configure search by the field `Account.CPF__c`

We also user this value to create a reference in the composite request,
but SalesForce return an error in that case:
```
{
      "errorCode": "JSON_PARSER_ERROR",
      "message": "Provided referenceId ('newQueryCustomFieldAccount.CPF__c') must start with a letter or a number, and it can contain only letters, numbers and underscores ('_')."
}
```

## Goal
Skip any character that is not a letter, number or underscore

## Issue ticket number and link

<!-- If you have a Jira ticket paste a link to the ticket related.
Remember to link it as [JIRA-NUMBER](http://url-to-the-ticket) -->

## Context
<!-- You can add any contextual information -->

## How have you test it?
<!-- Sometimes is just a unit-test. But also could be a series of steps
to reproduce the issue that the PR is addressing. Try to be as
informative as you can! -->

## Checklist before requesting a review
<!-- Please check the items you've already done, this is to ensue
everything is ready to be reviewed by your peers also feel free to add
more checks if you need -->
- [x] I have performed a self-review of my code
- [ ] If it is a core feature, I have added thorough tests.
- [ ] Do we need to implement analytics?
- [ ] Will this be part of a product update? If yes, please write one
phrase about this update.