fix(sfc_data): fix the function that creates the account to throw error (#20)
## Description
SF app returns status 201 when the request was successfull.

## Goal
<!-- Why are you making this change? what are you trying to achieve? -->

## Issue ticket number and link

<!-- If you have a Jira ticket paste a link to the ticket related.
Remember to link it as [JIRA-NUMBER](http://url-to-the-ticket) -->

## Context
<!-- You can add any contextual information -->

## How have you test it?
<!-- Sometimes is just a unit-test. But also could be a series of steps
to reproduce the issue that the PR is addressing. Try to be as
informative as you can! -->

## Checklist before requesting a review
<!-- Please check the items you've already done, this is to ensue
everything is ready to be reviewed by your peers also feel free to add
more checks if you need -->
- [ ] I have performed a self-review of my code
- [ ] If it is a core feature, I have added thorough tests.
- [ ] Do we need to implement analytics?
- [ ] Will this be part of a product update? If yes, please write one
phrase about this update.